inline HttpUtil functionality
